switched to SubCommands
