+++ src/main/java/de/slimecloud/slimeball/main/SlimeBot.java
@@ -13,8 +13,8 @@ import de.slimecloud.slimeball.features.alerts.HolidayAlert;
 import de.slimecloud.slimeball.features.alerts.Spotify;
 import de.slimecloud.slimeball.features.alerts.SpotifyAlert;
 import de.slimecloud.slimeball.features.birthday.Birthday;
-import de.slimecloud.slimeball.features.birthday.commands.BirthdayCommand;
 import de.slimecloud.slimeball.features.birthday.BirthdayTable;
+import de.slimecloud.slimeball.features.birthday.commands.BirthdayCommand;
 import de.slimecloud.slimeball.features.fdmds.FdmdsCommand;
 import de.slimecloud.slimeball.features.general.BonkCommand;
 import de.slimecloud.slimeball.features.general.BulkAddRoleCommand;
